Only remove the file if it's a different file and not the same but only differing in case.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@27481 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/SVN/SVNStatusListCtrl.cpp
@@ -6204,7 +6204,8 @@ void CSVNStatusListCtrl::OnRepairMove()
     // fix this by moving entry2 back to entry1 first,
     // then do an svn-move from entry1 to entry2
     CPathUtils::MakeSureDirectoryPathExists(entry1->GetPath().GetContainingDirectory().GetWinPath());
-    entry1->GetPath().Delete(true);
+    if (!entry1->GetPath().IsEquivalentToWithoutCase(entry2->GetPath()))
+        entry1->GetPath().Delete(true);
     if (!MoveFileEx(entry2->GetPath().GetWinPath(), entry1->GetPath().GetWinPath(), MOVEFILE_REPLACE_EXISTING))
     {
         // if entry2 existed before, the flag is still in the CTSVNPath object.
